+++ contracts/registry/CHANGELOG.md
@@ -13,13 +13,16 @@ Usage:
 Change log entries are to be added to the Unreleased section. Example entry:
 
 * [#<PR-number>](https://github.com/umee-network/umee/pull/<PR-number>) <description>
-
+-->
 
 # CHANGELOG: Registry
 
 ## Unreleased
 
 ### Features
+
+- Added `sbt_update_token_references` into the registry and SBT trait.
+- Added `token_reference` event (`Nep393Event::TokenReference`).
 
 ### Breaking Changes